+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.29.0] - 2022-02-09
+### Added
+- Paging for event list.
+
+### Fixed
+- Solar power widget for portal page.
+
 ## [1.28.0] - 2022-02-02
 ### Added
 - Solar power to 8 hour dashboard widget.